pyproject.toml
[tool.poetry]
name = "etl-airflow"
version = "0.1.0"
description = ""
authors = ["Francisco Santos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.6.4"
requests = "^2.31.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
taskipy = "^1.12.2"
ruff = "^0.3.4"
blue = "^0.9.1"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
exclude = ['.venv']

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = 'ruff . && blue --check . --diff'
format = 'blue .  && isort .'